requirements/common.txt
Django==1.11.12
django-cors-headers==2.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-cors-middleware==1.3.1
requests==2.18.4
olefile==0.44
pytz==2017.2
sorl-thumbnail==12.4.1
django-filter==1.1.0
psycopg2==2.7.3
Pillow==5.1.0
